Conflicts:
	GameData/TestFlight/Plugins/TestFlight.dll
	GameData/TestFlight/Plugins/TestFlightAPI.dll
	GameData/TestFlight/Plugins/TestFlightCore.dll
	TestFlight.userprefs

+++ TestFlightFailure_Engine.cs
@@ -157,7 +157,7 @@ namespace TestFlight
         {
             base.DoFailure();
         }
-        public override double DoRepair()
+        public override float DoRepair()
         {
             return base.DoRepair();
         }
